+++ CHANGELOG.md
@@ -21,6 +21,8 @@ Changelog
 
 **Next release**
 
+* Fixed applying template with dangerous function restrictions
+
 [Open issues](https://github.com/M66B/XPrivacy/issues?state=open)
 
 **Version 2.1.3 STABLE**